Move allocators.ts to services.ts, meaning.ts to utilities.ts, and transpile functions to a new file transpile.ts
