+++ src/services/tsconfig.json
@@ -42,7 +42,6 @@
         "../compiler/diagnosticInformationMap.generated.ts",
         "types.ts",
         "utilities.ts",
-        "allocators.ts",
         "breakpoints.ts",
         "classifier.ts",
         "completions.ts",
@@ -51,7 +50,6 @@
         "goToDefinition.ts",
         "jsDoc.ts",
         "jsTyping.ts",
-        "meaning.ts",
         "navigateTo.ts",
         "navigationBar.ts",
         "outliningElementsCollector.ts",
@@ -59,6 +57,7 @@
         "preProcess.ts",
         "rename.ts",
         "services.ts",
+        "transpile.ts",
         "shims.ts",
         "signatureHelp.ts",
         "symbolDisplay.ts",
